interaction_code/ElementImage/elementdetector/requirements.txt
Appium-Python-Client
pytest
selenium
pytest-ordering
pytest-html
Pillow
